Added test for testSortCommentsByTime
I added a comment but if you don't see it, it was "It should not be Time
but rather Date as that is what we have been using. " Another thing is
that Josh told us to call it "SortView" rather than "SortController" so
I changed it on this method. Up for debate/